Settings and Startup
On startup, if ChroniQuill finds it has already found a Home directory, it'll default to it and get you right to work.
If on startup, ChroniQuill finds an invalid settings.json in the Home directory, it'll ask you to pick the Home directory (there will be more validation to come)
Much debug code quarantined behind #if DEBUG #endifs
NIcer icons, ChoniQuill as CamelCase standarized throughout the code, except where all-lower-case is used.